pyproject.toml
[tool.poetry]
name = "syncflow"
version = "0.1.0"
description = ""
authors = ["tiw <tika11124@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
django-cors-headers = "^4.4.0"
django-filter = "^24.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
